Code clean-up
git-svn-id: https://svn.apache.org/repos/asf/openmeetings/trunk/singlewebapp@1621768 13f79535-47bb-0310-9956-ffa450edef68

+++ openmeetings-core/src/main/java/org/apache/openmeetings/core/session/SessionManager.java
@@ -291,7 +291,7 @@ public class SessionManager implements ISessionManager {
 				int start, int max, String orderby, boolean asc) {
 			SearchResult<Client> sResult = new SearchResult<Client>();
 			sResult.setObjectName(Client.class.getName());
-			sResult.setRecords(Long.valueOf(cache.size()).longValue());
+			sResult.setRecords(Long.valueOf(cache.size()));
 			sResult.setResult(cache.getClientsWithServer());
 			return sResult;
 		}
